requirements.txt
argparse==1.2.1
ase==3.11.0
bokeh==0.12.15
Cython==0.20.1
dateutils==0.6.6
Django==1.8.18
django-crispy-forms
django-extensions==1.2.5
django-filter==1.0.4
django-google-analytics==1
djangorestframework==3.6.4
djangorestframework-filters==0.11.1
djangorestframework-queryfields==1.0.0
djangorestframework-xml==1.4.0
djangorestframework-yaml==1.0.3
docutils==0.11
dynamic-rest==1.8.1
google-api-python-client==1.5.3
ipython==1.1.0
Jinja2==2.7.2
lark-parser==0.7.1
lxml==3.6.4
markdown==3.1
MarkupSafe==0.19
matplotlib==2.2.4
MySQL-python==1.2.5
networkx==1.8.1
nose==1.3.0
numpy==1.16.4
pexpect==4.2.1
ptyprocess==0.5.1
PuLP==1.6.10
PyCifRW==4.3
Pygments==1.6
pygraphviz==1.5
pyparsing==2.4.2
python-dateutil==2.2
python-memcached==1.53
pytz==2016.7
PyYAML==5.1.2
requests==2.21.0
scipy==0.13.3
six==1.10.0
spglib==1.14.0.post23
Sphinx==1.2.2
tornado==5.1.1
wsgiref==0.1.2
